feat: [UF-270] success page toast 수정 #280
