+++ src/app/login/page.tsx
@@ -6,14 +6,13 @@ import { useEffect } from 'react';
 import { toast } from 'sonner';
 
 import { ICON_PATHS } from '@/constants/icons';
-import { useToastStore } from '@/hooks/useToastStore';
 import { Button } from '@/shared/ui';
+import { useToastStore } from '@/stores/useToastStore';
 
 const LoginPage = () => {
   const { message, clearMessage } = useToastStore();
 
   useEffect(() => {
-    console.log(message);
     if (message) {
       toast.error(message);
       clearMessage();
